--- src/main/java/io/github/dqualizer/dqlang/types/dam/RequestParameter.java
@@ -1,8 +1,0 @@
-package io.github.dqualizer.dqlang.types.dam;
-
-import java.util.List;
-
-public class RequestParameter {
-    public String name;
-    public List<Scenario> scenarios;
-}
