+++ src/main/java/io/github/dqualizer/dqlang/types/dam/architecture/RequestParameter.java
@@ -1,0 +1,10 @@
+package io.github.dqualizer.dqlang.types.dam.architecture;
+
+import io.github.dqualizer.dqlang.types.dam.domain.Scenario;
+
+import java.util.List;
+
+public class RequestParameter {
+    public String name;
+    public List<Scenario> scenarios;
+}
