+++ src/main/java/io/github/dqualizer/dqlang/types/dam/architecture/Response.java
@@ -1,4 +1,4 @@
-package io.github.dqualizer.dqlang.types.dam;
+package io.github.dqualizer.dqlang.types.dam.architecture;
 
 import lombok.Data;
