+++ src/main/java/io/github/dqualizer/dqlang/types/rqa/configuration/loadtest/LoadTestArtifact.java
@@ -1,7 +1,7 @@
 package io.github.dqualizer.dqlang.types.rqa.configuration.loadtest;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import io.github.dqualizer.dqlang.types.dam.Endpoint;
+import io.github.dqualizer.dqlang.types.dam.architecture.Endpoint;
 import io.github.dqualizer.dqlang.types.rqa.definition.Artifact;
 import io.github.dqualizer.dqlang.types.rqa.definition.loadtest.ResponseMeasures;
 import io.github.dqualizer.dqlang.types.rqa.definition.stimulus.Stimulus;
