+++ src/main/kotlin/io/github/dqualizer/dqlang/types/rqa/configuration/instrumentation/InstrumentType.kt
@@ -1,4 +1,4 @@
-package io.github.dqualizer.dqlang.types.dam.instrumentation
+package io.github.dqualizer.dqlang.types.rqa.configuration.instrumentation
 
 /**
  * Instrument Types. Based on the opentelmetry spec.
