+++ src/main/kotlin/io/github/dqualizer/dqlang/types/rqa/configuration/monitoring/MonitoringConfiguration.kt
@@ -1,6 +1,6 @@
 package io.github.dqualizer.dqlang.types.rqa.configuration.monitoring
 
-import io.github.dqualizer.dqlang.types.dam.instrumentation.InstrumentationFramework
+import io.github.dqualizer.dqlang.types.dam.architecture.InstrumentationFramework
 
 data class MonitoringConfiguration(
     val serviceMonitoringConfigurations: Collection<ServiceMonitoringConfiguration>,
