+++ src/main/kotlin/io/github/dqualizer/dqlang/types/rqa/configuration/monitoring/ServiceMonitoringConfiguration.kt
@@ -1,7 +1,7 @@
 package io.github.dqualizer.dqlang.types.rqa.configuration.monitoring
 
-import io.github.dqualizer.dqlang.types.dam.instrumentation.Instrument
-import io.github.dqualizer.dqlang.types.dam.instrumentation.InstrumentationFramework
+import io.github.dqualizer.dqlang.types.rqa.configuration.instrumentation.Instrument
+import io.github.dqualizer.dqlang.types.dam.architecture.InstrumentationFramework
 
 data class ServiceMonitoringConfiguration(
     val serviceID: String,
